24-hour received times set toggle and all times reflect 24 hours
There's literally a split second where the local time is rendered as 12 hour even if the incoming URL says it should be 24 hour, but on the interval it fixes itself. Not as elegant as I'd like but it works.  I think that's the first time I've used "a split second" to mean exactly that.